+++ src/lib.rs
@@ -1,6 +1,9 @@
 // Use `README.md` as documentation home page, to reduce duplication
 #![doc = include_str!("../README.md")]
 
+/// Simple domain transformations to make propagators more generic.
+pub mod views;
+
 mod model;
 mod props;
 mod vars;